Make the variable typing stricter.
The type must now be one of:
 - bool
 - int or int32
 - int16
 - int64
 - float
 - double
 - string

WARNING: The variable type was previously not checked and using an unsupported type did not lead to any error message. Be aware that this change can make your existing configuration files invalid. However the adjustments required to adapt existing files should be minor.

git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS/trunk@527 d613d254-d18a-4304-b296-5125d4c4d670