Add the ability to append data to existing output file(s).
By default existing file(s) will still be overwritten. Set the new file attribute "append" to true if you wish to append data to existing NetCDF file(s).

Note that the append mode is currently not supported when file splitting is used and that the structure of the output file cannot be changed.

git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS/trunk@528 d613d254-d18a-4304-b296-5125d4c4d670