Changes concerning reading of a grid.
With introduction of new grid attributes and relaxing certain conditions on naming of attribues and coordinates in revisions r1430 and r1435, the algorithm on determining the type of a domain needed to be adjusted.

Tests: all three types of domain with various additional dimensions (scalar and axis with/without value/label).



git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS/dev/XIOS_DEV_CMIP6@1445 d613d254-d18a-4304-b296-5125d4c4d670